requirements.txt
flask
aiortc
opencv-python-headless
ultralytics
av
flask-cors
pymongo
python-dotenv
boto3
flask-socketio